+++ src/components/UI/Display/Organizations/OrganizationListPage.tsx
@@ -4,7 +4,7 @@ import useDebounce from 'lib/hooks/useDebounce';
 import {
   OrganizationNameStatusMessage,
   validateOrganizationName,
-} from 'lib/organizationValidation';
+} from 'MAPI/organizations/utils';
 import PropTypes from 'prop-types';
 import React, { useMemo, useState } from 'react';
 import styled from 'styled-components';
